feat: 마이페이지 위시리스트에서 품절 상품 장바구니 추가 시 예외 처리 및 '재고 없음' 반환

+++ src/main/resources/static/js/user/myWish.js
@@ -15,40 +15,52 @@ let csrfHeader = $("meta[name='_csrf_header']").attr("content");
 let csrfToken = $("meta[name='_csrf']").attr("content");
 
 function removeWish(productId) {
-    $.ajax({
-            url: '/ajax/wish',
-            type: 'DELETE',
-            data: {'productId': productId},
-            beforeSend: function (xhr) {
-                xhr.setRequestHeader(csrfHeader, csrfToken);
-            },
-            success: function (result) {
-                if(result.status == 200) {
-                    window.location.reload();
+    Swal.fire({
+        text: "위시리스트에서 삭제하시겠습니까?",
+        showCancelButton: true,
+        cancelButtonText: '아니요',
+        confirmButtonText: '예',
+        customClass: mySwalConfirm,
+        reverseButtons: true,
+        buttonsStyling: false,
+    }).then((result) => {
+        if (result.isConfirmed) {
+            $.ajax({
+                    url: '/ajax/wish',
+                    type: 'DELETE',
+                    data: {'productId': productId},
+                    beforeSend: function (xhr) {
+                        xhr.setRequestHeader(csrfHeader, csrfToken);
+                    },
+                    success: function (result) {
+                        if (result.status == 200) {
+                            window.location.reload();
+                        }
+                    },
+                    error: function () {
+                        if (xhr.status == 404) {
+                            Swal.fire({
+                                html: xhr.message.replace('\n', '<br>'),
+                                showConfirmButton: true,
+                                confirmButtonText: '확인',
+                                customClass: mySwal,
+                                buttonsStyling: false
+                            });
+                        } else {
+                            Swal.fire({
+                                html: "요청사항 진행 중 오류가 발생했습니다.<br>다시 시도해 주세요.",
+                                showConfirmButton: true,
+                                confirmButtonText: '확인',
+                                customClass: mySwal,
+                                buttonsStyling: false
+                            });
+                        }
+                    }
                 }
-            },
-            error: function () {
-                if(xhr.status == 404) {
-                    Swal.fire({
-                        html: xhr.message.replace('\n', '<br>'),
-                        showConfirmButton: true,
-                        confirmButtonText: '확인',
-                        customClass: mySwal,
-                        buttonsStyling: false
-                    });
-                } else {
-                    Swal.fire({
-                        html: "요청사항 진행 중 오류가 발생했습니다.<br>다시 시도해 주세요.",
-                        showConfirmButton: true,
-                        confirmButtonText: '확인',
-                        customClass: mySwal,
-                        buttonsStyling: false
-                    });
-                }
-            }
+            )
         }
-    )
-};
+    });
+}
 
 function addCart(productId) {
     $.ajax({
@@ -68,14 +80,25 @@ function addCart(productId) {
                 buttonsStyling: false
             });
         },
-        error: function () {
-            Swal.fire({
-                html: '장바구니에 삼품을 담는 중 오류가 발생했습니다.<br>다시 시도해 주세요.',
-                showConfirmButton: true,
-                confirmButtonText: '확인',
-                customClass: mySwal,
-                buttonsStyling: false
-            });
+        error: function (xhr) {
+            let result = JSON.parse(xhr.responseText);
+            if(xhr.status == 400) {
+                Swal.fire({
+                    html: result.message.replace('\n', '<br>'),
+                    showConfirmButton: true,
+                    confirmButtonText: '확인',
+                    customClass: mySwal,
+                    buttonsStyling: false
+                });
+            } else {
+                Swal.fire({
+                    html: '장바구니에 삼품을 담는 중 오류가 발생했습니다.<br>다시 시도해 주세요.',
+                    showConfirmButton: true,
+                    confirmButtonText: '확인',
+                    customClass: mySwal,
+                    buttonsStyling: false
+                });
+            }
         }
     })
 };
